pyproject.toml
[project]
name = "voice-spoofing-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kaggle>=2.0.1",
    "librosa>=0.11.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "soundfile>=0.13.1",
    "streamlit>=1.56.0",
    "streamlit-audiorecorder>=0.0.6",
    "torch>=2.11.0",
]
